requirements.txt
customtkinter
Pillow
torch
torchvision
pywinstyles
